May have fixed a bug

+++ db/migrate/20140528221637_remove_type_from_pages.rb
@@ -1,5 +1,5 @@
 class RemoveTypeFromPages < ActiveRecord::Migration
   def change
-		remove_column :pages, :type
+		remove_column :pages, :type if Page.new.respond_to?(:type)
   end
 end
